Adds additional tests and coverage support

+++ water/atexit_handler.py
@@ -6,26 +6,35 @@ from .display import display_debug
 
 CHILD_PIDS = set()
 
+
 def add_pid(pid: int):
     global CHILD_PIDS
     CHILD_PIDS.add(pid)
 
+
 def remove_pid(pid: int):
     global CHILD_PIDS
     if pid in CHILD_PIDS:
-      CHILD_PIDS.remove(pid)
+        CHILD_PIDS.remove(pid)
+
+
+def kill_pid(pid: int, sig: str = None, group: bool = True) -> None:
+    try:
+        pgid = os.getpgid(pid)
+        display_debug('sending', sig,
+                      f'{getattr(signal, sig)}', f'pgid={pgid}', f'pid={pid}')
+        if group:
+            os.killpg(os.getpgid(pid), getattr(signal, sig))
+        os.kill(pid, getattr(signal, sig))
+    except:
+        pass
+
 
 @atexit.register
 def cleanup_pids():
     global CHILD_PIDS
-    sig = 'SIGTERM'
     for pid in CHILD_PIDS:
-      try:
-        pgid = os.getpgid(pid)
-        display_debug('sending', sig, f'{getattr(signal, sig)}', f'pgid={pgid}', f'pid={pid}')
-        os.killpg(os.getpgid(pid), getattr(signal, sig))
-        os.kill(pid, getattr(signal, sig))
-      except:
-        pass
+        kill_pid(pid, sig='SIGTERM')
 
-__all__ = ['add_pid', 'remove_pid']
+
+__all__ = ['add_pid', 'remove_pid', 'kill_pid']
